A little refactoring and added triggered playback of the last action; still need to loop
