+++ src/FencingReplay/FencingReplay/App.xaml.cs
@@ -22,6 +22,8 @@ namespace FencingReplay
     /// </summary>
     sealed partial class App : Application
     {
+        internal FencingReplayConfig Config { get; set; }
+
         /// <summary>
         /// Initializes the singleton application object.  This is the first line of authored code
         /// executed, and as such is the logical equivalent of main() or WinMain().
@@ -30,6 +32,8 @@ namespace FencingReplay
         {
             this.InitializeComponent();
             this.Suspending += OnSuspending;
+
+            Config = FencingReplayConfig.FromFile();
         }
 
         /// <summary>
