+++ src/FencingReplay/FencingReplay/MainPage.xaml.cs
@@ -1,24 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Dynamic;
 using System.IO;
 using System.Linq;
 using System.Runtime.InteropServices.WindowsRuntime;
 using System.Threading.Tasks;
-using Windows.Foundation;
 using Windows.Foundation.Collections;
-using Windows.Graphics.Display;
-using Windows.Media.Capture;
 using Windows.Media.Capture.Frames;
-using Windows.Media.Core;
-using Windows.Media.MediaProperties;
 using Windows.Media.Playback;
 using Windows.Storage;
 using Windows.Storage.Pickers;
-using Windows.Storage.Streams;
-using Windows.System.Display;
-using Windows.UI.Core;
-using Windows.UI.Popups;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Controls.Primitives;
@@ -37,211 +27,12 @@ namespace FencingReplay
     /// </summary>
     public sealed partial class MainPage : Page
     {
-        class VideoChannel
-        {
-            MainPage mainPage;
-            int gridColumn;
-            ListBox sourceSelector;
-            CaptureElement captureElement;
-            MediaFrameSourceGroup currentSource;
-            MediaCapture currentCapture;
-            MediaPlayerElement mediaPlayerElement;
-            DisplayRequest currentRequest;
-            LowLagMediaRecording mediaRecording;
-            InMemoryRandomAccessStream currentRecordingStream;
-
-            bool isPreviewing = false;
-
-            static IReadOnlyList<MediaFrameSourceGroup> CurrentSources;
-
-            internal VideoChannel(MainPage page)
-            {
-                mainPage = page;
-
-                gridColumn = mainPage.LayoutGrid.ColumnDefinitions.Count;
-                var columnWidth = 300;
-                foreach (var cd in mainPage.LayoutGrid.ColumnDefinitions)
-                {
-                    cd.MaxWidth = columnWidth;
-                }
-                mainPage.LayoutGrid.ColumnDefinitions.Add(new ColumnDefinition() { MaxWidth = columnWidth });
-
-                sourceSelector = new ListBox();
-                PopulateSourceList(sourceSelector);
-                mainPage.LayoutGrid.Children.Add(sourceSelector);
-                Grid.SetColumn(sourceSelector, gridColumn);
-                Grid.SetRow(sourceSelector, 0);
-
-                captureElement = new CaptureElement();
-                mainPage.LayoutGrid.Children.Add(captureElement);
-                Grid.SetColumn(captureElement, gridColumn);
-                Grid.SetRow(captureElement, 1);
-
-                mediaPlayerElement = new MediaPlayerElement();
-                mediaPlayerElement.Visibility = Visibility.Collapsed;
-                mainPage.LayoutGrid.Children.Add(mediaPlayerElement);
-                Grid.SetColumn(mediaPlayerElement, gridColumn);
-                Grid.SetRow(mediaPlayerElement, 1);
-
-                Action<object, SelectionChangedEventArgs> eventHandler = (object sender, SelectionChangedEventArgs e) => SourceSelector_SelectionChanged(this, sender, e);
-                sourceSelector.SelectionChanged += new SelectionChangedEventHandler(eventHandler);
-
-                currentRequest = new DisplayRequest();
-            }
-
-            internal async Task Pause()
-            {
-            }
-
-            internal async Task Resume()
-            {
-
-            }
-
-            internal async Task<IAsyncAction> StartRecording(string fileBaseName)
-            {
-                currentRecordingStream = new InMemoryRandomAccessStream();
-                //var myVideos = await Windows.Storage.StorageLibrary.GetLibraryAsync(Windows.Storage.KnownLibraryId.Videos);
-                //StorageFile file = await myVideos.SaveFolder.CreateFileAsync($"{fileBaseName}-{gridColumn}.mp4", CreationCollisionOption.GenerateUniqueName);
-                //mediaRecording = await currentCapture.PrepareLowLagRecordToStorageFileAsync(MediaEncodingProfile.CreateMp4(VideoEncodingQuality.Auto), file);
-                mediaRecording = await currentCapture.PrepareLowLagRecordToStreamAsync(MediaEncodingProfile.CreateMp4(VideoEncodingQuality.Auto), currentRecordingStream);
-                return mediaRecording.StartAsync();
-            }
-
-            internal async Task<IAsyncAction> StopRecording()
-            {
-                await mediaRecording.StopAsync();
-                return mediaRecording.FinishAsync();
-            }
-
-            internal void StartPlayback()
-            {
-                captureElement.Visibility = Visibility.Collapsed;
-                mediaPlayerElement.Visibility = Visibility.Visible;
-                mediaPlayerElement.Source = MediaSource.CreateFromStream(currentRecordingStream, MediaEncodingProfile.CreateMp4(VideoEncodingQuality.Auto).ToString());
-                mediaPlayerElement.MediaPlayer.Play();
-            }
-
-            private async void SourceSelector_SelectionChanged(VideoChannel channel, object sender, SelectionChangedEventArgs e)
-            {
-                var source = FindMediaSource(channel.sourceSelector.SelectedItem.ToString());
-                if (source != null)
-                {
-                    channel.SetSource(source);
-                }
-                else
-                {
-                    channel.ClearSource();
-                }
-            }
-
-            async void ClearSource()
-            {
-                if (currentCapture != null)
-                {
-                    if (isPreviewing)
-                    {
-                        await currentCapture.StopPreviewAsync();
-                    }
-
-                    await mainPage.Dispatcher.RunAsync(CoreDispatcherPriority.Normal, () =>
-                    {
-                        captureElement.Source = null;
-                        if (currentRequest != null)
-                        {
-                            currentRequest.RequestRelease();
-                            currentRequest = new DisplayRequest();
-                        }
-
-                        currentCapture.Dispose();
-                        currentCapture = null;
-                        currentRecordingStream.Dispose();
-                        currentRecordingStream = null;
-                    });
-
-                    isPreviewing = false;
-                }
-            }
-
-            async void SetSource(MediaFrameSourceGroup sourceGroup)
-            {
-                ClearSource();
-
-                currentSource = sourceGroup;
-
-                try
-                {
-                    currentCapture = new MediaCapture();
-                    var settings = new MediaCaptureInitializationSettings { VideoDeviceId = GetVideoDeviceId(currentSource) };
-                    await mainPage.Dispatcher.RunAsync(CoreDispatcherPriority.Normal, async () =>
-                    {
-                        await currentCapture.InitializeAsync(settings);
-                        currentRequest.RequestActive();
-                        DisplayInformation.AutoRotationPreferences = DisplayOrientations.Landscape;
-                        captureElement.Source = currentCapture;
-                        await currentCapture.StartPreviewAsync();
-                        isPreviewing = true;
-                        mainPage.Paused = true;
-                    });
-
-                }
-                catch (UnauthorizedAccessException)
-                {
-                    // This will be thrown if the user denied access to the camera in privacy settings
-                    await (new MessageDialog("The app was denied access to the camera")).ShowAsync();
-                    return;
-                }
-                catch (System.IO.FileLoadException)
-                {
-                    //currentCapture.CaptureDeviceExclusiveControlStatusChanged += _mediaCapture_CaptureDeviceExclusiveControlStatusChanged;
-                }
-            }
-
-            static MediaFrameSourceGroup FindMediaSource(string displayName)
-            {
-                // This method will never get called when CurrentSources is uninitialized
-                foreach (var frameSource in CurrentSources)
-                {
-                    if (displayName == frameSource.DisplayName)
-                    {
-                        return frameSource;
-                    }
-                }
-                return null;
-            }
-
-            static string GetVideoDeviceId(MediaFrameSourceGroup sourceGroup)
-            {
-                foreach (var sourceInfo in sourceGroup.SourceInfos)
-                {
-                    if (sourceInfo.MediaStreamType == MediaStreamType.VideoRecord)
-                    {
-                        return sourceInfo.DeviceInformation.Id;
-                    }
-                }
-                return "";
-            }
-
-            static async void PopulateSourceList(ListBox listBox)
-            {
-                if (CurrentSources == null)
-                {
-                    CurrentSources = await MediaFrameSourceGroup.FindAllAsync();
-                }
-
-                listBox.Items.Clear();
-                foreach (var frameSource in CurrentSources)
-                {
-                    listBox.Items.Add(frameSource.DisplayName);
-                }
-                listBox.Items.Add("--none--");
-            }
-        }
-
         List<VideoChannel> channels;
 
         public bool Paused { get; set; }
         public bool Recording { get; set; }
+
+        private FencingReplayConfig config;
 
         public MainPage()
         {
@@ -253,6 +44,8 @@ namespace FencingReplay
 
             Paused = false;
             Recording = false;
+
+            config = (Application.Current as App).Config;
         }
 
         protected override void OnNavigatedTo(NavigationEventArgs e)
@@ -291,6 +84,7 @@ namespace FencingReplay
 
             PauseBtn.IsEnabled = true;
             PlayBtn.IsEnabled = true;
+            TriggerBtn.IsEnabled = true;
             Recording = true;
         }
 
@@ -306,6 +100,7 @@ namespace FencingReplay
 
             PauseBtn.IsEnabled = false;
             PlayBtn.IsEnabled = true;
+            TriggerBtn.IsEnabled = false;
             Recording = false;
         }
 
@@ -343,6 +138,22 @@ namespace FencingReplay
             foreach (var channel in channels)
             {
                 channel.StartPlayback();
+            }
+        }
+
+        private async void OnTrigger(object sender, RoutedEventArgs e)
+        {
+            if (config.ReplaySecondsAfterTrigger > 0)
+            {
+                await Task.Delay(1000 * config.ReplaySecondsAfterTrigger);
+                foreach (var channel in channels)
+                {
+                    await channel.StopRecording();
+                }
+                foreach (var channel in channels)
+                {
+                    channel.StartLoop(config.ReplaySecondsAfterTrigger + config.ReplaySecondsBeforeTrigger);
+                }
             }
         }
     }
